解决了跨平台导致的字节序不一致问题，python用大端，cpp平台是小端。 目前可以发送图像，返回坐标并解析。 可喜可贺

+++ portDemo/PointsPacket.h
@@ -3,9 +3,10 @@
 #include <vector>
 #include <string>
 #include <utility> // for std::pair
+#include <cstdint>
 
 // 坐标点数据包结构，服务端发送的数据包将会解析为这个格式
 struct PointsPacket {
 	std::string imageId;       // 图像ID
-	std::vector<std::pair<float, float>> points; // 坐标点数据
+	std::vector<std::pair<int32_t, int32_t>> points; // 坐标点数据
 };
